ai-coding-agents/opencode/requirements.txt
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
python-dotenv>=1.0.0
